+++ src/main/java/com/proyectoElSaber/domain/Estudiante.java
@@ -13,7 +13,7 @@ private static final long serialVersionUID = 1L;
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_estudiante")
-    private int idEstudiante;
+    private Long idEstudiante;
     private String seccion;  
     private String subirFoto;  
     private String nombreCompleto;
